Moved dragging code to separate class
Cleaned up card class and put draggable code in on script to make it
easy to make any gameobject draggable
