+++ Assets/Scripts/Card.cs
@@ -1,8 +1,8 @@
 ﻿using UnityEngine;
 using System.Collections;
-using UnityEngine.EventSystems;
 
-public class Card : MonoBehaviour, IBeginDragHandler, IDragHandler, IEndDragHandler {
+
+public class Card : MonoBehaviour{
 
     // Use this for initialization
     void Start() {
@@ -14,22 +14,4 @@ public class Card : MonoBehaviour, IBeginDragHandler, IDragHandler, IEndDragHand
 
     }
 
-    private Vector2 grabPoint;
-    public void OnBeginDrag(PointerEventData eventData)
-    {
-        //Keep the card "held" at the same position during the drag instead of snapping to the center
-        float diffX = this.transform.position.x - eventData.position.x;
-        float diffY = this.transform.position.y - eventData.position.y;
-        grabPoint = new Vector2(diffX, diffY);
-    }
-
-    public void OnDrag(PointerEventData eventData)
-    {
-        this.transform.position = eventData.position + grabPoint;
-    }
-
-    public void OnEndDrag(PointerEventData eventData)
-    {
-        Debug.Log("END DRAG");
-    }
 }
