+++ Assets/Scripts/Draggable.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a6c1cf35b2343d54ba1e140e9fd0ec5d
+timeCreated: 1466281728
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
